Pulled GraphQLValue out to its own file
